[vds3] fixes for glob results and wonky s3 prefixes
- Move `refresh` kwarg from `glob()` to `stat()`. That kwarg was
removed from `glob()` upstream, and _should_ be OK to remove here too.
If at some point we need interactive globbing and certainty that we're
getting the most recent data, we can explicitly invalidate caches. I'm
not inclined to mess with that unless it's necessary.

- Only abort opening an S3 path if both `fs.isdir()` and
`fs.exists()` fail. One or the other can show false for some S3
prefixes. (We keep trying to pretend this thing is a filesystem, and it
keeps not being a filesystem!) :)