[vds3] pull version information per key
Fetch version information individually for each object in a listing
sheet. The impact on performance and API call count for this change
depends on the structure of the bucket (flat vs. nested) and the number
of versions per object. We can make arguments in favor of either
approach, but until dask/s3fs#375 is more widely deployed this is a
safer option. Prior to that PR, listing more than 1,000 object
versions under a given prefix throws a pagination error.