Avoid shadowing `BaseSheet.refresh`
Rename `S3DirSheet.refresh()` to `refresh_path()`. This avoids shadowing
`BaseSheet.refresh()` which is related to refreshing the screen.
`refresh_path()` clears cached S3 object listings under a path.

If we intend to redraw the screen and instead reload the sheet data, it
is a lot of unnecessary work and can cause problems.

Closes #15 - multiple refreshes were launching async `reload()` threads
that clashed with each other and added multiple sets of rows and/or
columns.