[vd_jmespath] use a partial function for new jmespath columns
For expression columns, VisiData needs to eval the given expression as a
string. JMESPath expressions can have arbitrary levels of nested quotes,
so using them directly as column expressions doesn't work very well.

We already use partial functions in `select_by_jmespath`, and we can do
something similar here by stashing a partial function as a sheet
attribute and referencing that from the column expression.

This approach feels a little ugly, but it'll work more consistently than
the previous method in the case of nested quotes. And I'm not sure of a
better alternative.