pip-tools --> poetry
I was quite happy with pip-tools, but wrangling botocore got to be
untenable. I couldn't find a way to provide pip-tools enough hints
to find the "right" version of botocore given conflictin requests
from awscli, boto3 and aiobotocore (via s3fs). Poetry is able to get
there as long as we use tilde requirements for awscli and boto3.