Add vd_toml plugin (#21)
* Update vd_toml for submission as a dlc plugin

Punting to PyobjSheet was a _bit_ of a copout, so be a little more
explicit. Try to follow conventions of a more typical loader in the
hopes that this will be included in VisiData's dlc repo.

* pip-tools --> poetry

I was quite happy with pip-tools, but wrangling botocore got to be
untenable. I couldn't find a way to provide pip-tools enough hints
to find the "right" version of botocore given conflictin requests
from awscli, boto3 and aiobotocore (via s3fs). Poetry is able to get
there as long as we use tilde requirements for awscli and boto3.

* Add vd_toml tests